Move index logic to lib.tool_shed.util.shed_index
Since we need the index for browsing repos in the new client
we should incrementally build the index with every repo update.
Having this in tool_shed will make re-use easier.
